Renamed modules and package from maximum supported version to least supported version
