+++ custom/build.gradle.kts
@@ -3,8 +3,8 @@ plugins {
 }
 
 val instrumentations = listOf<String>(
-  ":instrumentation:openai-client-instrumentation:instrumentation-0.13.0",
-  ":instrumentation:openai-client-instrumentation:instrumentation-latest"
+  ":instrumentation:openai-client-instrumentation:instrumentation-0.2",
+  ":instrumentation:openai-client-instrumentation:instrumentation-0.14"
 )
 
 dependencies {
