+++ settings.gradle.kts
@@ -19,8 +19,8 @@ include("custom")
 include("instrumentation")
 include("instrumentation:openai-client-instrumentation:common")
 include("instrumentation:openai-client-instrumentation:testing-common")
-include("instrumentation:openai-client-instrumentation:instrumentation-0.13.0")
-include("instrumentation:openai-client-instrumentation:instrumentation-latest")
+include("instrumentation:openai-client-instrumentation:instrumentation-0.2")
+include("instrumentation:openai-client-instrumentation:instrumentation-0.14")
 include("inferred-spans")
 include("resources")
 include("smoke-tests")
